+++ broker/src/integration-test/groovy/com/swisscom/cloud/sb/broker/util/RestTemplateBuilderTest.groovy
@@ -21,11 +21,13 @@ import org.springframework.http.HttpEntity
 import org.springframework.http.HttpMethod
 import org.springframework.http.HttpStatus
 import org.springframework.web.client.RestTemplate
+import spock.lang.Ignore
 import spock.lang.Specification
 
+@Ignore
 class RestTemplateBuilderTest extends Specification {
-    private static final int http_port = 35000
-    private static final int https_port = 35001
+    private static final int http_port = 36000
+    private static final int https_port = 36001
 
 
     def "restTemplate with no features enabled"() {