+++ broker/src/main/groovy/com/swisscom/cloud/sb/broker/cfapi/dto/PlanDto.groovy
@@ -26,6 +26,7 @@ class PlanDto {
     String name
     String description
     boolean free
+    boolean active = true
     Map<String, Object> metadata
     List<ParameterDto> containerParams
     @JsonInclude(JsonInclude.Include.NON_NULL)