+++ broker/src/main/groovy/com/swisscom/cloud/sb/broker/cfextensions/endpoint/EndpointLookup.groovy
@@ -19,11 +19,9 @@ import com.swisscom.cloud.sb.broker.model.ServiceInstance
 import com.swisscom.cloud.sb.broker.util.servicedetail.ServiceDetailType
 import com.swisscom.cloud.sb.broker.util.servicedetail.ServiceDetailsHelper
 import com.swisscom.cloud.sb.model.endpoint.Endpoint
-import groovy.transform.CompileStatic
 import org.springframework.stereotype.Component
 
 @Component
-@CompileStatic
 class EndpointLookup {
     public static final Collection<String> DEFAULT_PROTOCOLS = Collections.unmodifiableCollection(["tcp"])